+++ src/rpc/replication.rs
@@ -32,6 +32,7 @@
 //! because otherwise the instance would not know if it should apply the DDL
 //! itself or wait for the tarantool replication.
 //!
+use crate::config::PicodataConfig;
 #[allow(unused_imports)]
 use crate::governor;
 #[allow(unused_imports)]
@@ -41,8 +42,7 @@ use crate::pico_service::pico_service_password;
 #[allow(unused_imports)]
 use crate::rpc;
 use crate::schema::PICO_SERVICE_USER_NAME;
-use crate::tarantool::box_ro_reason;
-use crate::tarantool::set_cfg_field;
+use crate::tarantool::{box_ro_reason, set_cfg_field, ListenConfig};
 use crate::tlog;
 use crate::traft::error::Error;
 #[allow(unused_imports)]
@@ -85,8 +85,11 @@ crate::define_rpc_request! {
 
         let mut replication_cfg = Vec::with_capacity(req.replicaset_peers.len());
         let password = pico_service_password();
+        let tls_config = &PicodataConfig::get().instance.iproto_tls;
         for address in &req.replicaset_peers {
-            replication_cfg.push(format!("{PICO_SERVICE_USER_NAME}:{password}@{address}"))
+            replication_cfg.push(ListenConfig::new(
+                format!("{PICO_SERVICE_USER_NAME}:{password}@{address}"),
+                tls_config));
         }
 
         crate::error_injection!("BROKEN_REPLICATION" => { replication_cfg.clear(); });